@@ -1,20 +1,30 @@
 {
   "name": "kdbush",
-  "version": "1.0.1",
+  "version": "3.0.0",
   "description": "A very fast static 2D index for points based on kd-tree.",
-  "main": "src/kdbush.js",
+  "module": "src/index.js",
+  "main": "kdbush.js",
+  "jsdelivr": "kdbush.min.js",
+  "unpkg": "kdbush.min.js",
   "repository": {
     "type": "git",
     "url": "git://github.com/mourner/kdbush.git"
   },
   "devDependencies": {
-    "eslint": "^3.8.1",
-    "eslint-config-mourner": "^2.0.1",
-    "tape": "^4.6.2"
+    "eslint": "^5.5.0",
+    "eslint-config-mourner": "^3.0.0",
+    "esm": "^3.0.82",
+    "rollup": "^0.65.2",
+    "rollup-plugin-buble": "^0.19.2",
+    "rollup-plugin-terser": "^2.0.2",
+    "tape": "^4.9.1"
   },
   "scripts": {
-    "pretest": "eslint test.js src",
-    "test": "tape test.js"
+    "pretest": "eslint src test.js bench.js rollup.config.js",
+    "test": "tape -r esm test.js",
+    "bench": "node -r esm bench.js",
+    "build": "rollup -c",
+    "prepublishOnly": "npm run build"
   },
   "eslintConfig": {
     "extends": "mourner"
@@ -27,6 +37,11 @@
     "algorithms",
     "spatial",
     "geometry"
+  ],
+  "files": [
+    "kdbush.js",
+    "kdbush.min.js",
+    "src"
   ],
   "author": "Vladimir Agafonkin",
   "license": "ISC"
