@@ -1,48 +1,50 @@
-'use strict';
 
-module.exports = sortKD;
-
-function sortKD(ids, coords, nodeSize, left, right, depth) {
+export default function sortKD(ids, coords, nodeSize, left, right, axis) {
     if (right - left <= nodeSize) return;
 
-    var m = Math.floor((left + right) / 2);
+    const m = (left + right) >> 1; // middle index
 
-    select(ids, coords, m, left, right, depth % 2);
+    // sort ids and coords around the middle index so that the halves lie
+    // either left/right or top/bottom correspondingly (taking turns)
+    select(ids, coords, m, left, right, axis);
 
-    sortKD(ids, coords, nodeSize, left, m - 1, depth + 1);
-    sortKD(ids, coords, nodeSize, m + 1, right, depth + 1);
+    // recursively kd-sort first half and second half on the opposite axis
+    sortKD(ids, coords, nodeSize, left, m - 1, 1 - axis);
+    sortKD(ids, coords, nodeSize, m + 1, right, 1 - axis);
 }
 
-function select(ids, coords, k, left, right, inc) {
+// custom Floyd-Rivest selection algorithm: sort ids and coords so that
+// [left..k-1] items are smaller than k-th item (on either x or y axis)
+function select(ids, coords, k, left, right, axis) {
 
     while (right > left) {
         if (right - left > 600) {
-            var n = right - left + 1;
-            var m = k - left + 1;
-            var z = Math.log(n);
-            var s = 0.5 * Math.exp(2 * z / 3);
-            var sd = 0.5 * Math.sqrt(z * s * (n - s) / n) * (m - n / 2 < 0 ? -1 : 1);
-            var newLeft = Math.max(left, Math.floor(k - m * s / n + sd));
-            var newRight = Math.min(right, Math.floor(k + (n - m) * s / n + sd));
-            select(ids, coords, k, newLeft, newRight, inc);
+            const n = right - left + 1;
+            const m = k - left + 1;
+            const z = Math.log(n);
+            const s = 0.5 * Math.exp(2 * z / 3);
+            const sd = 0.5 * Math.sqrt(z * s * (n - s) / n) * (m - n / 2 < 0 ? -1 : 1);
+            const newLeft = Math.max(left, Math.floor(k - m * s / n + sd));
+            const newRight = Math.min(right, Math.floor(k + (n - m) * s / n + sd));
+            select(ids, coords, k, newLeft, newRight, axis);
         }
 
-        var t = coords[2 * k + inc];
-        var i = left;
-        var j = right;
+        const t = coords[2 * k + axis];
+        let i = left;
+        let j = right;
 
         swapItem(ids, coords, left, k);
-        if (coords[2 * right + inc] > t) swapItem(ids, coords, left, right);
+        if (coords[2 * right + axis] > t) swapItem(ids, coords, left, right);
 
         while (i < j) {
             swapItem(ids, coords, i, j);
             i++;
             j--;
-            while (coords[2 * i + inc] < t) i++;
-            while (coords[2 * j + inc] > t) j--;
+            while (coords[2 * i + axis] < t) i++;
+            while (coords[2 * j + axis] > t) j--;
         }
 
-        if (coords[2 * left + inc] === t) swapItem(ids, coords, left, j);
+        if (coords[2 * left + axis] === t) swapItem(ids, coords, left, j);
         else {
             j++;
             swapItem(ids, coords, j, right);
@@ -60,7 +62,7 @@ function swapItem(ids, coords, i, j) {
 }
 
 function swap(arr, i, j) {
-    var tmp = arr[i];
+    const tmp = arr[i];
     arr[i] = arr[j];
     arr[j] = tmp;
 }
